+++ Lib/asyncio/__init__.py
@@ -16,6 +16,7 @@ from .streams import *
 from .subprocess import *
 from .tasks import *
 from .transports import *
+from .utils import *
 
 __all__ = (base_events.__all__ +
            coroutines.__all__ +
@@ -27,7 +28,8 @@ __all__ = (base_events.__all__ +
            streams.__all__ +
            subprocess.__all__ +
            tasks.__all__ +
-           transports.__all__)
+           transports.__all__ +
+           utils.__all__)
 
 if sys.platform == 'win32':  # pragma: no cover
     from .windows_events import *
