+++ Lib/asyncio/constants.py
@@ -6,5 +6,5 @@ ACCEPT_RETRY_DELAY = 1
 
 # Number of stack entries to capture in debug mode.
 # The larger the number, the slower the operation in debug mode
-# (see extract_stack() in events.py).
+# (see extract_stack() in utils.py).
 DEBUG_STACK_DEPTH = 10
