+++ Lib/asyncio/tasks.py
@@ -67,7 +67,8 @@ class Task(futures.Future):
         return {t for t in cls._all_tasks if t._loop is loop}
 
     def __init__(self, coro, *, loop=None):
-        assert coroutines.iscoroutine(coro), repr(coro)
+        if not coroutines.iscoroutine(coro):
+            raise TypeError(f"{coro!r} is not a coroutine")
         super().__init__(loop=loop)
         if self._source_traceback:
             del self._source_traceback[-1]
