+++ Lib/test/test_asyncio/utils.py
@@ -30,6 +30,7 @@ from asyncio import base_events
 from asyncio import events
 from asyncio import futures
 from asyncio import tasks
+from asyncio import utils
 from asyncio.log import logger
 from test import support
 
@@ -429,7 +430,7 @@ class MockPattern(str):
 
 
 def get_function_source(func):
-    source = events._get_function_source(func)
+    source = utils._get_function_source(func)
     if source is None:
         raise ValueError("unable to get the source of %r" % (func,))
     return source
